Fix clippy warnings for CI
- Add #[allow(dead_code)] to intentionally unused protocol engine types (ProtocolTriggerType, ProtocolCondition, ProtocolAction, Protocol, ProtocolContext, ProtocolResult, ProtocolEngine)
- Add #[allow(dead_code)] to unused Database fields (protocols Tree) and methods (search_substances, get_recent_food_logs)
- Add #[allow(dead_code)] to ReportSummary.days field
- Add #[allow(clippy::collapsible_if)] where nested ifs are intentional
- Add #[allow(unused_imports)] for tempdir in test files that need it for setup
- Remove unused imports (Commands, Duration, Utc, Uuid, fs, serde_yaml)
- Fix redundant closures and needless borrows in commands.rs
- Fix sort_by -> sort_by_key for efficiency
- Fix unnecessary cast in test
- Clean up push_str('\n') -> push('\n')

+++ tests/show_vitals_tests.rs
@@ -111,7 +111,7 @@ mod show_vitals_tests {
                 dbp: Some(70 + i),
                 temperature_c: Some(36.5 + i as f32 * 0.1),
                 spo2: Some(98),
-                hrv_rmssd: Some(40 + i as u32),
+                hrv_rmssd: Some(40 + i),
                 weight_kg: Some(70.0 + i as f32 * 0.5),
                 timestamp: Utc::now() - Duration::hours(i as i64),
                 notes: Some(format!("Log {}", i)),
